Restructure CLI: move CRUD verbs under domain subcommands, unify init fl
- Replace top-level `add`, `rm`, `ls` with `commands {new,rm,ls}` and `skills {new,rm,ls}`
- Replace four `--no-*` init flags with single `--features` whitelist flag
- Remove implicit `-v` → `--full` tie-in from `ls`
- Delete `src/cli/{add,rm,ls}.rs`; consolidate logic into `src/cli/commands.rs` and expanded `src/cli/skills.rs`
- Add OpenSpec proposal, design, specs, and tasks for change tracking
