+++ src/cli/mod.rs
@@ -1,10 +1,8 @@
-mod add;
+mod commands;
 mod completions;
 mod deploy;
 mod init;
-mod ls;
 mod options;
-mod rm;
 mod runner;
 mod skills;
 pub(crate) mod ui;
