+++ src/cli/runner.rs
@@ -1,17 +1,13 @@
-use super::add::run_add;
+use super::commands;
 use super::completions::generate_cli_completions;
 use super::deploy::deploy;
 use super::init::initialize_agents_dir;
-use super::ls::run_ls;
-use super::options::{Action, Options, SkillsAction};
-use super::rm::run_rm;
+use super::options::{Action, Options};
 use super::skills;
 use anyhow::Result;
 use clap::CommandFactory;
 
 pub(crate) fn run(opts: Options) -> Result<bool> {
-    let verbosity = opts.verbosity;
-
     let default_action = || {
         Options::command().print_help().unwrap();
         std::process::exit(0);
@@ -30,18 +26,8 @@ pub(crate) fn run(opts: Options) -> Result<bool> {
             deploy(opts)?;
             true
         }
-        Action::Skills { action } => match action {
-            SkillsAction::Add(opts) => skills::add(opts)?,
-        },
-        Action::Ls(mut opts) => {
-            // -v / --verbose global flag also enables full descriptions.
-            if verbosity > 0 {
-                opts.verbose = true;
-            }
-            run_ls(opts)?
-        }
-        Action::Add { action } => run_add(action)?,
-        Action::Rm { action } => run_rm(action)?,
+        Action::Skills { action } => skills::run_skills(action)?,
+        Action::Commands { action } => commands::run_commands(action)?,
     };
 
     Ok(success)
