+++ src/cli/skills.rs
@@ -1,15 +1,61 @@
-use anyhow::{Result, bail};
+use anyhow::{Context, Result, bail};
+use cliclack::{confirm, input, intro, outro};
+use gray_matter::Matter;
+use gray_matter::engine::YAML;
+use serde_json::Value;
+use std::fs;
 use std::io::ErrorKind;
 
+use crate::cli::deploy::deploy;
+use crate::cli::options::{
+    AddSkillOptions, DeployOptions, RmSkillOptions, SkillsAction, SkillsAddOptions, SubLsOptions,
+};
+use crate::cli::ui::ls::{ListItem, render_skills};
+use crate::constants::templates::{SKILL_STARTER, render_starter};
+use crate::prelude::*;
 use crate::schema::config::app::AppConfig;
 use crate::schema::config::common::PackageRunner;
 use crate::templates::get_templater;
-use crate::utils::path::get_application_dir;
+use crate::utils::fs::write_file;
+use crate::utils::path::{get_application_dir, get_skills_dir};
+use crate::utils::tty::is_tty;
 
-use super::options::SkillsAddOptions;
+/// Collect a string field: use provided value, prompt in TTY mode, or default to empty.
+fn collect_field(value: Option<String>, prompt: &str, placeholder: &str) -> Result<String> {
+    if let Some(v) = value {
+        return Ok(v);
+    }
+    if is_tty() {
+        let v: String = input(prompt)
+            .placeholder(placeholder)
+            .default_input("")
+            .interact()
+            .context(format!("failed to read {}", prompt))?;
+        return Ok(v);
+    }
+    Ok(String::new())
+}
+
+/// Prompt the user to deploy now (TTY only, used when --deploy was not passed).
+fn maybe_prompt_deploy(deploy_flag: bool) -> Result<()> {
+    if deploy_flag {
+        deploy(DeployOptions::default()).context("deploy failed")?;
+        return Ok(());
+    }
+    if is_tty() {
+        let should_deploy = confirm("Deploy now?")
+            .initial_value(false)
+            .interact()
+            .unwrap_or(false);
+        if should_deploy {
+            deploy(DeployOptions::default()).context("deploy failed")?;
+        }
+    }
+    Ok(())
+}
 
 /// Install a skill into `.dotagents/skills/` by wrapping the `skills` CLI.
-pub(crate) fn add(opts: SkillsAddOptions) -> Result<bool> {
+fn add(opts: SkillsAddOptions) -> Result<bool> {
     let templater = get_templater();
     let app_config = AppConfig::from_application(templater)?;
 
@@ -62,4 +108,199 @@ pub(crate) fn add(opts: SkillsAddOptions) -> Result<bool> {
         .status()?;
 
     Ok(status.success())
+}
+
+/// Handle `dotagents skills new`.
+fn new_skill(opts: AddSkillOptions) -> Result<bool> {
+    let app_dir = get_application_dir().context(
+        "No .dotagents directory found. Run `dotagents init` to initialise a workspace.",
+    )?;
+
+    let skill_dir = app_dir.join("skills").join(&opts.name);
+    let target = skill_dir.join("SKILL.md");
+
+    if target.exists() && !opts.force {
+        bail!(
+            "Skill '{}' already exists at {}. Use --force to overwrite.",
+            opts.name,
+            target.display()
+        );
+    }
+
+    let use_interactive = is_tty()
+        && opts.description.is_none()
+        && opts.license.is_none()
+        && opts.compatibility.is_none();
+
+    if use_interactive {
+        intro("dotagents skills new").ok();
+    }
+
+    let description = collect_field(opts.description, "Description", "What does this skill do?")?;
+    let license = collect_field(opts.license, "License", "e.g. MIT")?;
+    let compatibility = collect_field(
+        opts.compatibility,
+        "Compatibility",
+        "e.g. Requires openspec CLI.",
+    )?;
+
+    // Build frontmatter.
+    let mut frontmatter = format!("name: {}\n", opts.name);
+    frontmatter.push_str(&format!("description: \"{}\"\n", description));
+    if !license.is_empty() {
+        frontmatter.push_str(&format!("license: {}\n", license));
+    }
+    if !compatibility.is_empty() {
+        frontmatter.push_str(&format!("compatibility: \"{}\"\n", compatibility));
+    }
+    frontmatter.push_str("metadata:\n  version: \"1.0\"\n");
+
+    let body = render_starter(SKILL_STARTER, &opts.name);
+    let content = format!("---\n{}---\n\n{}", frontmatter, body);
+
+    write_file(&target, &content).context("failed to write SKILL.md")?;
+
+    success!("Created {}", target.display());
+
+    maybe_prompt_deploy(opts.deploy)?;
+
+    if use_interactive {
+        outro("").ok();
+    }
+
+    Ok(true)
+}
+
+/// Handle `dotagents skills rm`.
+fn rm_skill(opts: RmSkillOptions) -> Result<bool> {
+    let app_dir = get_application_dir().context(
+        "No .dotagents directory found. Run `dotagents init` to initialise a workspace.",
+    )?;
+
+    let skill_dir = app_dir.join("skills").join(&opts.name);
+
+    if !skill_dir.exists() {
+        bail!(
+            "Skill '{}' not found at {}.",
+            opts.name,
+            skill_dir.display()
+        );
+    }
+
+    intro("dotagents skills rm").ok();
+
+    // Confirm in TTY unless --force.
+    if is_tty() && !opts.force {
+        let confirmed = confirm(format!(
+            "Remove skill '{}'? This cannot be undone.",
+            opts.name
+        ))
+        .initial_value(false)
+        .interact()
+        .unwrap_or(false);
+
+        if !confirmed {
+            outro("Cancelled.").ok();
+            return Ok(true);
+        }
+    }
+
+    fs::remove_dir_all(&skill_dir)
+        .with_context(|| format!("failed to remove {}", skill_dir.display()))?;
+
+    success!("Removed {}", skill_dir.display());
+
+    maybe_prompt_deploy(opts.deploy)?;
+
+    outro("").ok();
+    Ok(true)
+}
+
+/// Load all skills from `.dotagents/skills/*/SKILL.md`, returning name+description pairs.
+fn load_skills() -> Result<Vec<ListItem>> {
+    let skills_dir = match get_skills_dir() {
+        Ok(d) => d,
+        Err(_) => return Ok(vec![]), // skills dir absent → empty
+    };
+
+    let matter = Matter::<YAML>::new();
+    let mut items = Vec::new();
+
+    for entry in fs::read_dir(&skills_dir).context("failed to read skills directory")? {
+        let entry = entry?;
+        let path = entry.path();
+        if !path.is_dir() {
+            continue;
+        }
+        let skill_md = path.join("SKILL.md");
+        if !skill_md.is_file() {
+            continue;
+        }
+        let content = fs::read_to_string(&skill_md).context("failed to read SKILL.md")?;
+        let Ok(parsed) = matter.parse::<Value>(&content) else {
+            continue;
+        };
+        if let Some(data) = parsed.data {
+            let name: String = data["name"].as_str().unwrap_or("").to_string();
+            let description: String = data["description"].as_str().unwrap_or("").to_string();
+            if !name.is_empty() {
+                items.push(ListItem { name, description });
+            }
+        }
+    }
+
+    items.sort_by(|a, b| a.name.cmp(&b.name));
+    Ok(items)
+}
+
+/// Handle `dotagents skills ls`.
+fn ls_skills(opts: SubLsOptions) -> Result<bool> {
+    get_application_dir().context(
+        "No .dotagents directory found. Run `dotagents init` to initialise a workspace.",
+    )?;
+    let skills = load_skills().context("failed to load skills")?;
+    render_skills(skills, opts.full);
+    Ok(true)
+}
+
+/// Dispatch `dotagents skills`.
+pub(crate) fn run_skills(action: SkillsAction) -> Result<bool> {
+    match action {
+        SkillsAction::Add(opts) => add(opts),
+        SkillsAction::New(opts) => new_skill(opts),
+        SkillsAction::Rm(opts) => rm_skill(opts),
+        SkillsAction::Ls(opts) => ls_skills(opts),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::fs;
+    use tempfile::TempDir;
+
+    fn make_skill(skills_dir: &std::path::Path, name: &str, description: &str) {
+        let skill_dir = skills_dir.join(name);
+        fs::create_dir_all(&skill_dir).unwrap();
+        let content = format!(
+            "---\nname: {}\ndescription: \"{}\"\n---\n\nBody.\n",
+            name, description
+        );
+        fs::write(skill_dir.join("SKILL.md"), content).unwrap();
+    }
+
+    #[test]
+    fn load_skills_reads_skill_md_frontmatter() {
+        // load_skills parses name and description from SKILL.md files
+        let tmp = TempDir::new().unwrap();
+        make_skill(tmp.path(), "my-skill", "Does something");
+
+        let matter = Matter::<YAML>::new();
+        let skill_md = tmp.path().join("my-skill").join("SKILL.md");
+        let content = fs::read_to_string(&skill_md).unwrap();
+        let parsed = matter.parse::<Value>(&content).unwrap();
+        let data = parsed.data.unwrap();
+        assert_eq!(data["name"].as_str().unwrap(), "my-skill");
+        assert_eq!(data["description"].as_str().unwrap(), "Does something");
+    }
 }
